pagong/lib/pdf_checker/requirements.txt
# PDF Checker 필수 라이브러리

# PDF 처리
PyMuPDF==1.23.8      # PDF 분석 및 텍스트 추출
pdf2image==1.16.3    # PDF를 이미지로 변환

# OCR
pytesseract==0.3.10  # Tesseract OCR Python 래퍼
easyocr==1.7.1       # 딥러닝 기반 OCR (한국어 지원 우수)
# 주의: Tesseract OCR 엔진을 별도로 설치해야 합니다
# macOS: brew install tesseract tesseract-lang
# Ubuntu: sudo apt-get install tesseract-ocr tesseract-ocr-kor

# 이미지 처리
Pillow==10.1.0       # 이미지 처리
opencv-python==4.8.1.78  # 이미지 품질 분석

# 텍스트 분석
python-Levenshtein==0.23.0  # 레벤슈타인 거리 계산
difflib              # Python 내장 라이브러리 (설치 불필요)

# 수치 계산
numpy==1.24.3        # 수치 연산

# 딥러닝 (EasyOCR 의존성)
torch>=1.7.0         # PyTorch (CPU 버전)
torchvision>=0.8.0   # 컴퓨터 비전 라이브러리

# 기타
poppler-utils        # pdf2image 의존성 (시스템 패키지)
# macOS: brew install poppler
# Ubuntu: sudo apt-get install poppler-utils